done commit

+++ pos/src/main/java/com/increff/pos/dto/UserDto.java
@@ -30,13 +30,29 @@ public class UserDto {
     public UserPojo get(String email) throws ApiException {
         return service.get(email);
     }
-
+    public UserPojo get(int id){
+        return service.get(id);
+    }
     public List<UserData> getAll() {
          return  conversion(service.getAll());
 
     }
+    public List<UserData> getLimited(Integer pageNo) {
+        return  conversion(service.getLimited(pageNo));
 
-    public void delete(int id) {
+    }
+    public Long getTotalNoBrands() {
+
+        return service.getTotalNoBrands();
+    }
+
+    public void delete(int id) throws ApiException {
+        UserPojo p = service.get(id);
+
+        if(p.getRole().equals("supervisor"))
+        {
+            throw new ApiException("You can't delete a Supervisor");
+        }
         service.delete(id);
     }
     public UserPojo validate(UserPojo p)
